requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
Pillow
numpy
python-multipart